Add test for insufficient gas at DAO split
This test makes sure that if there is not enough gas given at a
daoSplit() event the new DAO is not created at all.

This should work thanks to this homestead change:
https://github.com/ethereum/EIPs/blob/master/EIPS/eip-2.mediawiki#specification

Before homestead the test would fail, but create an empty/invalid new
DAO address which would lead to people burning their tokens in their old
DAO and transferring them to a new invalid DAO, thus losing their money.